Add transactions/{hash}/state-changes endpoint; State changes code refactoring
