+++ apps/block_scout_web/lib/block_scout_web/api_router.ex
@@ -114,6 +114,7 @@ defmodule BlockScoutWeb.ApiRouter do
       get("/:transaction_hash/internal-transactions", V2.TransactionController, :internal_transactions)
       get("/:transaction_hash/logs", V2.TransactionController, :logs)
       get("/:transaction_hash/raw-trace", V2.TransactionController, :raw_trace)
+      get("/:transaction_hash/state-changes", V2.TransactionController, :state_changes)
     end
 
     scope "/blocks" do
